MapChildren function placed in TreeView and renamed
