+++ examples/src/Treeview.tsx
@@ -17,6 +17,7 @@ export class Index extends React.Component<any, any> {
             treeviewElements: elements
         };
     }
+
     public render() {
         return (
             <div>
